requirements.txt
requests==2.31.0
pytest==7.4.0
allure-pytest==2.13.2
pytest-ordering==0.6